+++ src/app/components/popover-category/popover-category.page.spec.ts
@@ -2,6 +2,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { PopoverCategoryPage } from './popover-category.page';
+import {RouterTestingModule} from '@angular/router/testing';
 
 describe('PopoverCategoryPage', () => {
   let component: PopoverCategoryPage;
@@ -10,7 +11,7 @@ describe('PopoverCategoryPage', () => {
   beforeEach(async(() => {
     TestBed.configureTestingModule({
       declarations: [ PopoverCategoryPage ],
-      imports: [IonicModule.forRoot()]
+      imports: [IonicModule.forRoot(), RouterTestingModule]
     }).compileComponents();
 
     fixture = TestBed.createComponent(PopoverCategoryPage);
