+++ src/services/todo/todo.service.spec.ts
@@ -1,12 +1,17 @@
 import { TestBed } from '@angular/core/testing';
 
 import { TodoService } from './todo.service';
+import {RouterTestingModule} from '@angular/router/testing';
+import {ModalController} from '@ionic/angular';
 
 describe('TodoService', () => {
   let service: TodoService;
 
   beforeEach(() => {
-    TestBed.configureTestingModule({});
+    TestBed.configureTestingModule({
+      imports: [RouterTestingModule],
+      providers: [ModalController]
+    });
     service = TestBed.inject(TodoService);
   });
